Move Input to pkg/model
